requirements.txt
mysql-connector-python~=8.0.25
PyQt5~=5.15.4